% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/port_plot.R
\name{EfficientFrontier}
\alias{EfficientFrontier}
\title{Plot Efficient frontier}
\usage{
EfficientFrontier(
  asset.names,
  increment = 100,
  rf = 0,
  period = c("months", "weeks", "quarters", "years")
)
}
\arguments{
\item{asset.names}{Vector of ticker of securities}

\item{increment}{Number of portfolio to be generated, Default: 100}

\item{rf}{Risk-free rate of return, Default: 0}

\item{period}{Period for which the returns are calculated, Default: c("months", "weeks", "quarters", "years")}
}
\value{
Interactive plot of the effiecnt frontier for given security/funds
}
\description{
Plots the efficient frontier for the security (mean-variance portfolios)
}
\details{
Give and interactive effient frontier for given securities
}
\examples{
EfficientFrontier(c('FXAIX', 'TIBFX'), period = 'days')
}
