% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize_cone_chart.R
\name{generate_cone_chart}
\alias{generate_cone_chart}
\title{ConeChart for the asset or fund under investigation
Evaluate the returns of an investment manager giventheir return and volatility
expectations. Calculates 1 and 2 standard deviations from the expected return.}
\usage{
generate_cone_chart(
  return_series,
  target_return,
  expected_volatility,
  timeseries_frequency
)
}
\arguments{
\item{return_series}{: The return series of the asset or fund that you are planning to invest in}

\item{target_return}{: The target return you have for the asset or fund that you are planning to invest in}

\item{expected_volatility:}{The expected volatility you have for the asset or fund that you are planning to invest in}

\item{timeseries_frequency:}{Daily, Monthly, Weekly, Yearly}
}
\value{

}
\description{
ConeChart for the asset or fund under investigation
Evaluate the returns of an investment manager giventheir return and volatility
expectations. Calculates 1 and 2 standard deviations from the expected return.
}
\examples{
Not Run
}
