% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_summary_stats.R
\name{generate_select_summary_statistics}
\alias{generate_select_summary_statistics}
\title{Generate A Table Of Selected Summary Statistics for Potential Investments}
\usage{
generate_select_summary_statistics(return_series, target_returns, rf_rate)
}
\arguments{
\item{return_series}{: The return series of the asset or fund that you are planning to invest in}

\item{target_return}{: The target return you have for the asset or fund that you are planning to invest in}

\item{rf_rate:}{The rf rate on your investment}
}
\value{

}
\description{
Generate A Table Of Selected Summary Statistics for Potential Investments
}
\examples{
# Not Run generate_select_summary_statis
}
