% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/portfolio_vol_contribution.R
\name{portfolio_vol_contribution}
\alias{portfolio_vol_contribution}
\title{Calculate and Graph "Percent contribution to volatility"}
\usage{
portfolio_vol_contribution(weights, portfolioComponentReturns)
}
\arguments{
\item{weights:}{A numeric vector indicating the weight of the investment in asset}

\item{portfolioComponentReturns:}{an xts of portfolio returns - ensure that}
}
\value{

}
\description{
Calculate and Graph "Percent contribution to volatility"
}
\examples{
Not Run
}
